prevent even more warnings
clang-format rtt sources

+++ lib/rtt.c
@@ -1,7 +1,6 @@
-/* include rtt1 */
-#include    "unprtt.h"
+#include "unprtt.h"
 
-int     rtt_d_flag = 0;     /* debug flag; can be set by caller */
+int rtt_d_flag = 0; /* debug flag; can be set by caller */
 
 /*
  * Calculate the RTO value based on current estimators:
@@ -9,46 +8,43 @@ int     rtt_d_flag = 0;     /* debug flag; can be set by caller */
  */
 #define RTT_RTOCALC(ptr) ((ptr)->rtt_srtt + (4.0 * (ptr)->rtt_rttvar))
 
-static float
-rtt_minmax(float rto) {
+static double
+rtt_minmax(double rto) {
     if (rto < RTT_RXTMIN) {
         rto = RTT_RXTMIN;
     } else if (rto > RTT_RXTMAX) {
         rto = RTT_RXTMAX;
     }
-    return(rto);
+    return (rto);
 }
 
 void
 rtt_init(struct rtt_info *ptr) {
-    struct timeval  tv;
+    struct timeval tv;
 
     Gettimeofday(&tv, NULL);
-    ptr->rtt_base = tv.tv_sec;      /* # sec since 1/1/1970 at start */
+    ptr->rtt_base = tv.tv_sec; /* # sec since 1/1/1970 at start */
 
     ptr->rtt_rtt    = 0;
     ptr->rtt_srtt   = 0;
     ptr->rtt_rttvar = 0.75;
-    ptr->rtt_rto = rtt_minmax(RTT_RTOCALC(ptr));
+    ptr->rtt_rto    = rtt_minmax(RTT_RTOCALC(ptr));
     /* first RTO at (srtt + (4 * rttvar)) = 3 seconds */
 }
-/* end rtt1 */
 
 /*
  * Return the current timestamp.
  * Our timestamps are 32-bit integers that count milliseconds since
  * rtt_init() was called.
  */
-
-/* include rtt_ts */
 uint32_t
 rtt_ts(struct rtt_info *ptr) {
-    uint32_t        ts;
-    struct timeval  tv;
+    uint32_t       ts;
+    struct timeval tv;
 
     Gettimeofday(&tv, NULL);
     ts = ((tv.tv_sec - ptr->rtt_base) * 1000) + (tv.tv_usec / 1000);
-    return(ts);
+    return (ts);
 }
 
 void
@@ -58,10 +54,9 @@ rtt_newpack(struct rtt_info *ptr) {
 
 int
 rtt_start(struct rtt_info *ptr) {
-    return((int)(ptr->rtt_rto + 0.5));      /* round float to int */
-    /* 4return value can be used as: alarm(rtt_start(&foo)) */
+    return ((int)(ptr->rtt_rto + 0.5)); /* round double to int */
+    /* return value can be used as: alarm(rtt_start(&foo)) */
 }
-/* end rtt_ts */
 
 /*
  * A response was received.
@@ -71,13 +66,11 @@ rtt_start(struct rtt_info *ptr) {
  * This function should be called right after turning off the
  * timer with alarm(0), or right after a timeout occurs.
  */
-
-/* include rtt_stop */
 void
 rtt_stop(struct rtt_info *ptr, uint32_t ms) {
-    double      delta;
+    double delta;
 
-    ptr->rtt_rtt = ms / 1000.0;     /* measured RTT in seconds */
+    ptr->rtt_rtt = ms / 1000.0; /* measured RTT in seconds */
 
     /*
      * Update our estimators of RTT and mean deviation of RTT.
@@ -86,39 +79,34 @@ rtt_stop(struct rtt_info *ptr, uint32_t ms) {
      */
 
     delta = ptr->rtt_rtt - ptr->rtt_srtt;
-    ptr->rtt_srtt += delta / 8;     /* g = 1/8 */
+    ptr->rtt_srtt += delta / 8; /* g = 1/8 */
 
     if (delta < 0.0) {
-        delta = -delta;    /* |delta| */
+        delta = -delta; /* |delta| */
     }
 
-    ptr->rtt_rttvar += (delta - ptr->rtt_rttvar) / 4;   /* h = 1/4 */
+    ptr->rtt_rttvar += (delta - ptr->rtt_rttvar) / 4; /* h = 1/4 */
 
     ptr->rtt_rto = rtt_minmax(RTT_RTOCALC(ptr));
 }
-/* end rtt_stop */
 
 /*
  * A timeout has occurred.
  * Return -1 if it's time to give up, else return 0.
  */
-
-/* include rtt_timeout */
 int
 rtt_timeout(struct rtt_info *ptr) {
-    ptr->rtt_rto *= 2;      /* next RTO */
+    ptr->rtt_rto *= 2; /* next RTO */
 
     if (++ptr->rtt_nrexmt > RTT_MAXNREXMT) {
-        return(-1);    /* time to give up for this packet */
+        return (-1); /* time to give up for this packet */
     }
-    return(0);
+    return (0);
 }
-/* end rtt_timeout */
 
 /*
  * Print debugging information on stderr, if the "rtt_d_flag" is nonzero.
  */
-
 void
 rtt_debug(struct rtt_info *ptr) {
     if (rtt_d_flag == 0) {
